+++ Launcher/App/src/main/java/com/tobbe/pilauncher/MainActivityNear.java
@@ -1,0 +1,4 @@
+package com.tobbe.pilauncher;
+
+public class MainActivityNear extends MainActivity {
+}