+++ CHANGELOG.md
@@ -1,6 +1,16 @@
-# Latest version: [PiLauncher_2.0.apk](https://github.com/Tobbe85/PiLauncher/releases/download/2.0/PiLauncher_2.0.apk)
+# Latest version: [PiLauncher_2.1.apk](https://github.com/Tobbe85/PiLauncher/releases/download/2.1/PiLauncher_2.1.apk)
 
 ## Changelog
 
-**v2.0**
-- Initial Release
+**v2.1**
+- Now the Default Style matches the Pico Style on Pico
+- Engine Text got an own row in Detail App View
+- Improved Settings Symbol in Main View
+- In PiLauncher Settings you can now choose the Mode (Docked / Windowed) on Pico
+- Updates "Available Storage" Text directly after deleting an app
+- Click on the "Available Storage" Info opens the Storage Manager
+- Added a Battery Indicator on bottom row
+- Long Press on Main View activates or deactivates "Edit Mode"
+- Added an own WebView instead of open the Browser for Game Information
+- Start Animation by clicking on an app (Zoom & Fade)
+- Icons will be stored in /data/data/com.tobbe.pilauncher/files now